feat: add service layer test and improve factory

+++ internal/application/expense/service.go
@@ -2,13 +2,24 @@ package expense
 
 import (
 	"context"
+	"errors"
 	"expenses/internal/domain"
 
 	"github.com/shopspring/decimal"
 )
 
+var (
+	ErrInvalidExpenseID = errors.New("expense id must be positive")
+)
+
 type Service struct {
 	repository domain.Repository
+}
+
+type ExpenseSummary struct {
+	TotalAmount decimal.Decimal `json:"total_amount"`
+	TotalBudget decimal.Decimal `json:"total_budget"`
+	Difference  decimal.Decimal `json:"difference"`
 }
 
 func NewService(repository domain.Repository) *Service {
@@ -26,17 +37,20 @@ func (s *Service) GetAll(ctx context.Context, filter domain.Filter) ([]domain.Ex
 }
 
 func (s *Service) Delete(ctx context.Context, id int64) error {
+
+	if id <= 0 {
+		return ErrInvalidExpenseID
+	}
+
 	return s.repository.Delete(ctx, id)
 }
 
 func (s *Service) GetByID(ctx context.Context, id int64) (domain.Expense, error) {
-	return s.repository.GetByID(ctx, id)
-}
 
-type ExpenseSummary struct {
-	TotalAmount decimal.Decimal `json:"total_amount"`
-	TotalBudget decimal.Decimal `json:"total_budget"`
-	Difference  decimal.Decimal `json:"difference"`
+	if id <= 0 {
+		return domain.Expense{}, ErrInvalidExpenseID
+	}
+	return s.repository.GetByID(ctx, id)
 }
 
 func (s *Service) GetExpensesWithSummary(ctx context.Context) ([]domain.Expense, ExpenseSummary, error) {
